Changed TemperatureHelper conversion functions to handle Decimal type to avoid floating point arithmetic errors with Doubles.

+++ Dewy/Helpers/Temperature.swift
@@ -13,12 +13,12 @@ import Foundation
 struct TemperatureHelper {
     
     /// Converts a given celcius value into fahrenheit.
-    static func toFahrenheit(celcius: Double) -> Double {
+    static func toFahrenheit(celcius: Decimal) -> Decimal {
         return (celcius * 9/5) + 32
     }
     
     /// Converts a given fahrenheit value into celcius.
-    static func toCelcius(fahrenheit: Double) -> Double {
+    static func toCelcius(fahrenheit: Decimal) -> Decimal {
         return (fahrenheit - 32) * 5/9
     }
 }
